feat: 更新

+++ result/responseBean.go
@@ -1,7 +1,7 @@
 package result
 
 type ResponseSuccessBean struct {
-	Code uint32 `json:"code"`
+	Code int64  `json:"code"`
 	Msg  string `json:"msg"`
 	Data any    `json:"data,omitempty"`
 }
